Json refactor

+++ runtime/include/admissions_info.h
@@ -4,12 +4,12 @@
 
 struct admissions_info {
 	struct perf_window perf_window;
-	int                percentile;        /* 50 - 99 */
+	uint8_t            percentile;        /* 50 - 99 */
 	int                control_index;     /* Precomputed Lookup index when perf_window is full */
 	uint64_t           estimate;          /* cycles */
 	uint64_t           relative_deadline; /* Relative deadline in cycles. This is duplicated state */
 };
 
-void admissions_info_initialize(struct admissions_info *admissions_info, int percentile, uint64_t expected_execution,
-                                uint64_t relative_deadline);
+void admissions_info_initialize(struct admissions_info *admissions_info, uint8_t percentile,
+                                uint64_t expected_execution, uint64_t relative_deadline);
 void admissions_info_update(struct admissions_info *admissions_info, uint64_t execution_duration);
